requirements.txt
Django~=4.2.6
psutil~=5.9.6
paramiko~=3.3.1
python-nmap~=0.7.1
scapy~=2.5.0
pyinstaller~=6.0.0
requests~=2.31.0
WMI~=1.4.9
beautifulsoup4~=4.12.2
nginxparser~=1.0.1